Handle content task polling only on the manage content page.
