+++ kolibri/plugins/device/assets/src/views/ManageContentPage/index.vue
@@ -72,6 +72,8 @@
   import get from 'lodash/get';
   import sortBy from 'lodash/sortBy';
   import { mapState, mapGetters, mapActions } from 'vuex';
+  import { useIntervalFn } from '@vueuse/core';
+  import { getCurrentInstance } from 'kolibri.lib.vueCompositionApi';
   import commonCoreStrings from 'kolibri.coreVue.mixins.commonCoreStrings';
   import { TaskResource } from 'kolibri.resources';
   import taskNotificationMixin from '../taskNotificationMixin';
@@ -97,6 +99,25 @@
       TasksBar,
     },
     mixins: [commonCoreStrings, taskNotificationMixin],
+    setup() {
+      const polling = useIntervalFn(() => {
+        $store.dispatch('manageContent/refreshTaskList');
+      }, 1000);
+      const $store = getCurrentInstance().proxy.$store;
+      function startTaskPolling() {
+        if ($store.getters.canManageContent) {
+          polling.resume();
+        }
+      }
+      function stopTaskPolling() {
+        polling.pause();
+      }
+
+      return {
+        stopTaskPolling,
+        startTaskPolling,
+      };
+    },
     data() {
       return {
         deleteChannelId: null,
@@ -167,6 +188,12 @@
       if (!this.channelsAreInstalled) {
         this.$store.commit('SET_WELCOME_MODAL_VISIBLE', true);
       }
+    },
+    mounted() {
+      this.startTaskPolling();
+    },
+    destroyed() {
+      this.stopTaskPolling();
     },
     methods: {
       ...mapActions('manageContent', ['refreshChannelList', 'startImportWorkflow']),
